Refactor ctrl-c download test
* Move event waiter to init module so we can reuse across tests
* Modify the measured time to be from ctrl-c until exit.  We don't
  really care when the download() call was started, we care that
  once ctrl-c a transfer, it exits in a reasonable amount of time.
